requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-restful==0.3.10
flask-principal==0.4.0
chardet==5.2.0
celery==5.4.0
flower==2.0.1
flask-sqlalchemy==3.1.1
langid==1.1.6
networkx==3.2.1
babel==2.14.0
elasticsearch==7.17.12
elasticsearch-dsl==7.4.1
iso639==0.1.4
couchdb==1.2
pymongo==4.6.1
redis==5.0.1
lxml==5.0.0
zipstream==1.1.4
uwsgi==2.0.23
kytea==0.1.7
requests==2.31.0
requests-toolbelt==1.0.0
treetaggerwrapper==2.3
nltk==3.8.1
pyyaml==6.0.1
editdistance==0.6.2
translate==3.6.1
future==0.18.3
pytest>=7.4.0
pytest-html>=3.2.0
pytest-xdist>=3.3.0
psycopg2==2.9.9
pexpect==4.9.0
microsofttranslator==0.8
cmreshandler==1.0.0
theano==1.0.5
gunicorn==21.2.0
opensearch-dsl==2.1.0
opensearch-py==2.2.0
pyjwt==2.8.0
cryptography==41.0.7
amqp==5.2.0